refactor pdf compression
